requirements.txt
numpy
opencv-python
pathlib
Pillow
